✨ Simple networking for python functions
Only a few socket calls are implemented as we expect to implement more
on a need-basis. Currently, this wraps a "TCPStream" so it only
supports TCP and not UDP.
