+++ runtimes/python/examples/firm-api/setup.py
@@ -7,5 +7,9 @@ setup(
     author_email="pipeline@goodbyekansas.com",
     description="Example showcasing the Firm API in Python",
     py_modules=["firm_api"],
-    entry_points={"console_scripts": ["firm-api=firm_api:main"]},
+    entry_points={
+        "console_scripts": [
+            "firm-api=firm_api:main", "firm-api-error=firm_api:main_with_error"
+        ],
+    },
 )
